refactor(tauri): Parameters (-P) をフッター右端に移動
CommandForm 下部の fieldset を廃止し、Footer の固定バー右端に
BlueEditButton として配置。状態管理・隠しフィールド・ダイアログも Footer へ移動。
